Fixed bug that was preventing zipfiles from being uploaded in a directory is the archive or its contents exist in the users home directory.  Added try catch to properly propagate the upload zip errors.

+++ stochss/handlers/file_browser.py
@@ -308,35 +308,37 @@ class MoveFileAPIHandler(APIHandler):
         for proj_item in os.listdir(old_path,):
             if proj_item.endswith('.wkgp'):
                 for exp_item in os.listdir(os.path.join(old_path, proj_item)):
-                    with open(os.path.join(old_path, proj_item, exp_item,
-                                           "info.json"), "r+") as info_file:
-                        info = json.load(info_file)
-                        log.debug("Old wkfl info: %s", info)
-                        if get_status(os.path.join(old_path, proj_item,
-                                                   exp_item)) == 'ready':
-                            if not old_parent_dir:
-                                src_mdl = os.path.join(new_parent_dir,
-                                                       info['source_model'])
+                    wkfl_path = os.path.join(old_path, proj_item, exp_item)
+                    if wkfl_path.endswith(".wkfl"):
+                        with open(os.path.join(wkfl_path,
+                                               "info.json"), "r+") as info_file:
+                            info = json.load(info_file)
+                            log.debug("Old wkfl info: %s", info)
+                            if get_status(os.path.join(old_path, proj_item,
+                                                       exp_item)) == 'ready':
+                                if not old_parent_dir:
+                                    src_mdl = os.path.join(new_parent_dir,
+                                                           info['source_model'])
+                                else:
+                                    src_mdl = (info['source_model']
+                                               .replace(old_parent_dir, new_parent_dir, 1))
+                                info['source_model'] = (src_mdl[1:]
+                                                        if src_mdl.startswith('/')
+                                                        else src_mdl)
                             else:
-                                src_mdl = (info['source_model']
-                                           .replace(old_parent_dir, new_parent_dir, 1))
-                            info['source_model'] = (src_mdl[1:]
-                                                    if src_mdl.startswith('/')
-                                                    else src_mdl)
-                        else:
-                            if not old_parent_dir:
-                                wkfl_mdl = os.path.join(new_parent_dir,
-                                                        info['wkfl_model'])
-                            else:
-                                wkfl_mdl = info['wkfl_model'].replace(old_parent_dir,
-                                                                      new_parent_dir, 1)
-                            info['wkfl_model'] = (wkfl_mdl[1:]
-                                                  if wkfl_mdl.startswith('/')
-                                                  else wkfl_mdl)
-                        info_file.seek(0)
-                        log.debug("New wkfl info: %s", info)
-                        json.dump(info, info_file)
-                        info_file.truncate()
+                                if not old_parent_dir:
+                                    wkfl_mdl = os.path.join(new_parent_dir,
+                                                            info['wkfl_model'])
+                                else:
+                                    wkfl_mdl = info['wkfl_model'].replace(old_parent_dir,
+                                                                          new_parent_dir, 1)
+                                info['wkfl_model'] = (wkfl_mdl[1:]
+                                                      if wkfl_mdl.startswith('/')
+                                                      else wkfl_mdl)
+                            info_file.seek(0)
+                            log.debug("New wkfl info: %s", info)
+                            json.dump(info, info_file)
+                            info_file.truncate()
 
 
 class DuplicateModelHandler(APIHandler):
@@ -737,9 +739,20 @@ class UploadFileAPIHandler(APIHandler):
         file_name = file_data['filename']
         log.debug(file_name)
         log.debug(type(file_data['body']))
-        resp = upload(file_data, file_info)
-        log.debug(resp)
-        self.write(json.dumps(resp))
+        try:
+            resp = upload(file_data, file_info)
+            log.debug(resp)
+            self.write(json.dumps(resp))
+        except StochSSAPIError as err:
+            self.set_status(err.status_code)
+            error = {"Reason":err.reason, "Message":err.message}
+            if err.traceback is None:
+                trace = traceback.format_exc()
+            else:
+                trace = err.traceback
+            log.error("Exception information: %s\n%s", error, trace)
+            error['Traceback'] = trace
+            self.write(error)
         self.finish()
